demo for pip install dry-run, pip cache
Performs dry-run on a list of packages and show cache information.

Currently, pip does not support listing the cache content, because the cache operates on the HTTP responses. Since the original HTTP request URL was not stored alongside the response, the cached files cannot be easily associated with the packages. Future pip behavior is subject to change.

https://github.com/pypa/pip/issues/10460